backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
slowapi>=0.1.9
